Fixed SkyHanni gui edit button not working inside storage even if neu storage is not disabled.
