+++ src/main/java/at/hannibal2/skyhanni/features/misc/SkyBlockKickDuration.kt
@@ -24,8 +24,20 @@ class SkyBlockKickDuration {
     fun onChat(event: LorenzChatEvent) {
         if (!isEnabled()) return
         if (event.message == "§cYou were kicked while joining that server!") {
-            //if is in hub: enable rn
-            kickMessage = true
+
+            if (LorenzUtils.onHypixel && !LorenzUtils.inSkyBlock) {
+                kickMessage = false
+                showTime = true
+                lastKickTime = SimpleTimeMark.farPast()
+            } else {
+                kickMessage = true
+            }
+        }
+
+        if (event.message == "§cThere was a problem joining SkyBlock, try again in a moment!") {
+            kickMessage = false
+            showTime = true
+            lastKickTime = SimpleTimeMark.farPast()
         }
     }
